Move rollDice to utils
